Change storage for now & last to uint32_t
millis() will wrap in a uint16_t in a little over 1 minute

+++ examples/StatusMonitor/StatusMonitor.ino
@@ -15,8 +15,8 @@ BasicTerm term(&Serial);
 
 #define LED 13
 
-int16_t now;
-int16_t last;
+uint32_t now;
+uint32_t last;
 
 void setup(void) {
     Serial.begin(9600);
